requirements.txt
# Pure Momentum Trading System Requirements
# Optimized for CRCL-type explosive move detection

# Core ML and Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=1.7.0
scipy>=1.11.0

# Async HTTP and WebSocket
aiohttp>=3.8.0
websockets>=11.0.0
asyncio-mqtt>=0.11.0

# API and Data
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
alpaca-py>=0.29.0
alpaca-trade-api>=3.0.0

# Date/Time handling
python-dateutil>=2.8.0
pytz>=2023.3

# System and Monitoring
psutil>=5.9.0
logging>=0.4.9.6

# Optional: Performance optimization
numba>=0.57.0